test(dart_source_generator): move tests to folder to match package structure
