+++ packages/globe_runtime_ts/dart_source_generator/test/src/type-mapper.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { mapTsTypeToDart } from "../src/type-mapper";
+import { mapTsTypeToDart } from "../../src/type-mapper";
 import ts from "typescript";
 
 function createTypeNode(code: string): ts.TypeNode {
