+++ packages/globe_runtime_ts/dart_source_generator/test/src/utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { toCamelCase } from "../src/utils";
+import { toCamelCase } from "../../src/utils";
 
 describe("utils", () => {
   it("should convert snake_case to camelCase", () => {
